Add Transformers and Controllers
Add FraudMobile, FraudAccount, FraudEmails Transformers and their respective Controllers

+++ database/migrations/2017_10_09_100005_create_fraud_cases_table.php
@@ -19,7 +19,7 @@ class CreateFraudCasesTable extends Migration
             $table->dateTime('scam_start_date')->nullable();
             $table->dateTime('scam_realization_date')->nullable();
             $table->integer('severity_id')->unsigned();
-            $table->decimal('amount_scammed_off', 8, 2)->unsigned();
+            $table->double('amount_scammed_off', 10, 2)->unsigned();
             $table->integer('fraud_category_id')->unsigned()->index();
             $table->string('scammer_name');
             $table->string('scammer_real_name')->nullable();
